+++ draft-packages/menu/KaizenDraft/Menu/components/MenuList.tsx
@@ -1,6 +1,6 @@
-import * as React from "react"
+import React from "react"
 import { v4 } from "uuid"
-import styles from "../styles.scss"
+import styles from "./MenuList.module.scss"
 
 type MenuListProps = {
   heading?: React.ReactNode