+++ draft-packages/hero-card/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [2.1.16](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.1.15...@kaizen/draft-hero-card@2.1.16) (2022-07-27)
+
+**Note:** Version bump only for package @kaizen/draft-hero-card
+
+
+
+
+
 ## [2.1.15](https://github.com/cultureamp/kaizen-design-system/compare/@kaizen/draft-hero-card@2.1.14...@kaizen/draft-hero-card@2.1.15) (2022-07-27)
 
 **Note:** Version bump only for package @kaizen/draft-hero-card